% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwlmm.R, R/rgwlmm.R
\name{gwlmm}
\alias{gwlmm}
\alias{rgwlmm}
\title{(Robust) Geographically weighted linear mixed model (GWLMM)}
\usage{
gwlmm(formula, data, band = NULL, centroid = FALSE, maxit = 100,
  tol = 1e-04, ...)

rgwlmm(formula, data, band = NULL, centroid = FALSE, maxit = 100,
  tol = 1e-04, k = 1.345, Start = NULL, method = "Fix", ...)
}
\arguments{
\item{formula}{(formula) a two-sided lineer formula object decribing
the fixed effects, the random intercept and the geographical information of the model.
The response is on the left of the ~ operator, and the x-variables on the right side
are seperated by a + opeartor. The identifier for the random intercept is seperated
with a vertical bar (|).After another (|), the two coordinates - longitude and latitude -
are seperated by a + operator.
Categorial x-variable need to be defined as factor-variables.}

\item{data}{(data.frame) A dataframe containing the variables named in \code{formula}.}

\item{band}{(numeric) A numeric value defining the bandwidth for the geographical weigths (default = NULL).
For a predifined bandwdth, insert value here.}

\item{centroid}{(logical) If coordinates in \code{formula} are constant within the area ID define cetroid = TRUE (default = FALSE)}

\item{maxit}{(integer) Defines the maximum number of iterations for the fitting process (default = 100).}

\item{tol}{(numeric) Defines the tolerance for the convergence of the fitting process (default = 1e-04).}

\item{...}{not used}

\item{k}{(numeric) defines the tuning constant for influience function (default = 1.346). Ses datails.}

\item{Start}{(list) optioanl list containing three obejcts for the starting
values for the robust approximation (default = NULL).
The three objects are: \code{betas} (Matrix with local coefficients);
\code{sigma.v} (numeric value for the variance of the random effects);
\code{sigma.e} (numeric value for the error term variance).}

\item{method}{(character) defines the iterative algorithm for approximating the variance
parameters. Possible values: \code{"Fix", "Newton"}, (default = \code{"Fix"}). See datails.}
}
\value{
The function \code{gwlmm} returns an object of class \code{gwlmm}. The function \code{rgwlmm} returns
an object of class \code{rgwlmm}. Both objects are lists containing the following elements
\itemize{
\item \code{call} (language) the call generating the value
\item \code{coefficients} (Matrix) the matrix of local regression coefficients
\item \code{varcoefficients} (Matrix) the matrix containing the
model-based variances of the local coefficients
\item \code{Variance} (numeric) the vector of fitted variance parameter(s)
\item \code{nIter} (numeric) number of iterations needed
for estimating the model parameters
\item \code{LRtest} (numeric) a named vector of test results for spatial
non-stationarity conating the likelihood, the likelihood ratio,
 the effective degrees of freedom and the p-value. (only in \code{gwlmm}-object)
\item \code{Model} (list) contains all model information for forther calculations
}
}
\description{
User interface for fitting a GWLMM model. The function \code{gwlmm} fits a GWLMM via REML
whereas \code{rgwlmm} fits an outlier-robust GWLMM to data. The GWLMM is a random
intercespt model that take into account spatial non-stationarity in the model coefficients.
}
\details{
\itemize{
\item \code{gwlmm} additionally conducts a likelihood ratio  using the likelihood from the GWLMM
and a global LMM. The LMM is fitted using the \code{lmer} function from the \code{lme4}-package.
The test is saved in the \code{LRtest} value and is based on Fotheringham et al. (2009, p. 92).

\item The bandwidth is optimized using the cross-validation procedure implemented
 in \code{gwr.sel} from the \code{spgwr}-package.

\item The REML estimation implemented in the \code{gwlmm} function is conducted using the Nelder-Mead algirithm (\code{optim} from \code{stats}-package)
as suggested by Chandra et al. (2012).
}

\itemize{
\item  \code{rgwlmm} uses robust ML estimation equations (following Sinha and Rao, 2009)
of the GWLMM that restrict the infuence of outliers on the parameter estimation using
Hubers's influence function (See Baldermann et al. (2016)). The tuning constant \code{k} defines the restriction. The smaller \code{k}, the stronger
the restriction and vice versa.Sinha and Rao (2009) recommended to set \code{k = 1.346}
following Huber (1964).
\item  \code{method}: using a fixed-point algorithm is recommended for estimating the variance parameters as it
is fast and stable compared to the Newton-Raphsom algorithm.

}
}
\examples{
# Data sets ?sampleData, ?popaggData and ?popoutData are
# implemented in the rsarGWR-package. See help files.

\dontrun{

formula <- y~1+x|clusterid |long + lat

#Model fit
gwmodel<-gwlmm(formula, data = data)

#In-sample predictions
pred<-predict(gwmodel)

#Small area mean prediction for aggregated population data
predagg<-predict(gwmodel, popdata = popaggData, size = "Size")

#Small area mean prediction for unit-level population data
preddisagg<-predict(gwmodel, popdata = popoutData, popAgg = FALSE)
}

##################################################################
# Outlier-robust estimation
\dontrun{

# Model fit
rgwmodel<- rgwlmm(formula, data = sampleData)

# In-sample prediction
rpred<-predict(rgwmodel)
#Small area preditions (mean) for aggregated population data
rpredagg<-predict(rgwmodel, popdata = popaggData, size = "Size")
#Small area preditions (mean) for unit-level population data
rpreddisagg<-predict(rgwmodel, popdata = popoutData, popAgg = FALSE)

###########

# Robust model fit when sample only contains centroid information
rgwmodel<- rgwlmm(formula, data = sampleData, centroid = TRUE)

# In-sample prediction
rpred<-predict(rgwmodel)
#Small area means for aggregated population data
rpredagg<-predict(rgwmodel, popdata = popaggData, size = "Size")

}

}
\references{
Baldermann, C., N. Salvati, T. Schmid (2016). Robust small area estimation under spatial
non-stationarity. Working Paper.

Chandra, H., N. Salvati, R. Chambers, and N. Tzavidis (2012). Small area
estimation under spatial nonstationarity. Computational Statistics and Data Analysis 56 (10),
pp. 2875-2888.

Fotheringham, A. S., C. Brunsdon, and M. Charlton (2002). Geographically Weighted Regression. West
Sussex: Wiley.

Huber, P. J. (1964). Robust estimation of a location parameter. The Annals of Mathematical Statistical 35,
73 - 101.

Bivand, R and D. Yu (2015). spgwr: Geographically Weighted Regression. R package
version 0.6-28. https://CRAN.R-project.org/package=spgw

Sinha, S. K. and J. N. K. Rao (2009). Robust small area estimation. The Canadian Journal of Statistics 37
(3), 381 - 399.
}
\seealso{
\code{\link{predict.gwlmm}}, and \code{\link{predict.rgwlmm}}
}

