% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{popaggData}
\alias{popaggData}
\title{Agrregated population for demonstrational purpose of the rsaeGWR package}
\format{A data frame with 4 rows and 6 variable:
\describe{
  \item{clusterid}{ID variable, defining the areas}
  \item{long}{Longituge of the centroid}
  \item{lat}{Latitude of the centroid}
  \item{x}{Continious explanatory variable}
  \item{Size}{Area-specific population size}
}}
\usage{
data(popaggData)
}
\description{
Agrregated population for demonstrational purpose of the rsaeGWR package
}
\keyword{datasets}

