% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{popoutData}
\alias{popoutData}
\title{Generated out-of-sample population units}
\format{A data frame with 4000 rows and 6 variable:
\describe{
  \item{clusterid}{ID variable, defining the areas}
  \item{long}{Longituge }
  \item{lat}{Latitude}
  \item{x}{Continious explanatory variable}
  }}
\usage{
data(popoutData)
}
\description{
Generated out-of-sample population units
}
\keyword{datasets}

