% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwlmm.predict.R, R/rgwlmm.predict.R
\name{predict.gwlmm}
\alias{predict.gwlmm}
\alias{predict.rgwlmm}
\title{Predict Method for (Robust) GWLMM Fits}
\usage{
\method{predict}{gwlmm}(object, popdata = NULL, size = NULL,
  popAgg = TRUE, ...)

\method{predict}{rgwlmm}(object, popdata = NULL, size = NULL,
  popAgg = TRUE, bcConst = 3, maxit = 100, tol = 1e-04, ...)
}
\arguments{
\item{object}{(formula) object of class \code{'gwlmm'} or \code{'rgwlmm'}}

\item{popdata}{(data.frame) an optional data frame (Default = NULL). If ommited,
in-sample prdictions are estimated.}

\item{size}{(character) name for the column in \code{popdata}
containing the area-specific
population sizes (Default = NULL). Obligatory when \code{popAgg = TRUE}.}

\item{popAgg}{(logical) \code{TRUE} if popdata is aggregated and only contains
area level information}

\item{...}{not used}

\item{bcConst}{(numeric) needed for objects of class \code{rgwlmm}. Defines the tuning constant for influience function
in the bias correction term (default = 3). Ses datails.}

\item{maxit}{(integer) needed for objects of class \code{rgwlmm}. Defines the maximum number of iterations for
the estiamtion of the random effects for \code{rgwlmm} objects (default = 100).}

\item{tol}{(numeric) needed for objects of class \code{rgwlmm}. Defines the tolerance for the convergence of the fitting process (default = 1e-04).}
}
\value{
The function \code{predict.gwlmm} returns a list containing the predictions

If \code{popdata = NULL}, a list with the following elements is returned.

\itemize{
\item \code{call} (language) the call generating the value
\item \code{nIter} (numeric) number of iterations needed for the estimation of the
random effects (only for \code{rgwlmm}-objects)
\item \code{raneff} (numeric) named vector of random effects
\item \code{residuals} (numeric) vector of restimated residuals
\item \code{prediction} (numeric) vector of estimated in-sample predicted values
\item \code{xbeta} (numeric) vector of estimated fixed part of the in-sample predictions
}

If \code{popdata} is a data.frame a list with the following elements is returned.

\itemize{
\item \code{call} (language) the call generating the value
\item \code{nIter} (numeric) number of iterations needed for the estimation of the
random effects (only for \code{rgwlmm}-objects)
\item \code{areaEst} (data.frame) a data frame containing area-specific mean predictions and MSE estimates.
\item \code{ranEff} (data.farme) a data frame containing the random effects.
\item \code{sampleEst} (data.frame) a data frame containing the in-sample
predictions (prediction, residuals, xbeta)
}
}
\description{
Interface for predicting small area means based on a
\code{\link{gwlmm}} ( or \code{\link{rgwlmm}}) model fit.
}
\details{
\itemize{
\item The argument \code{popdata} can have three different definitions:
(1) \code{popdata = NULL},
(2) \code{popdata} is a data frame for aggregated population information,
(3) \code{popdata} is a data frame for unit-level population information.
In case (1) only in-sample predictions are estimated. In case (2) the \code{size} argument is obligatory.
In case (3) \code{popAgg} must be \code{TRUE}. Population data can contain non-sampled areas.

\item The method \code{predict} is implemented for three combinations of
geographic information in the sampled (S) and population (P)
data: (1) only centroid information in S and P; (2) unit-level geographic information in S and P;
(3)  unit-level geographic information in S and centriod informaton in P.

\item For objects of class \code{gwlmm}: the MSE estimates for the small area means returned in \code{areaEst} are named \code{CCT} and \code{CCST}. The former is based
on pseudo-linearization following Chambers et al. (2011), and latter is linearization-based following
Chambers et al. (2014).

\item For objects of class \code{rgwlmm}: the MSE estimates for the small area means returned in \code{areaEst}
are named \code{CCT} (and \code{CCT_bc})  and \code{CCST} ( and \code{CCST_bc}). The subscript \code{_bc} indicates the respective
MSE estimate for bias corrected robust area mean. The MSE estimates are based on the pseudo-linearizetion (CCT)
and the linearization-based approach in Chambers et al. (2014).

\item For objects of class \code{rgwlmm}: the bias correction is based on the area-specific prediction error where the influence of extreme
values is restricted using Huber's influence function. \code{cbConst} defines this restriction but
should be less restrictive the \code{k} in \code{\link{rgwlmm}}. By default it is set \code{cbConst = 3}.

}
}
\examples{
# Data sets ?sampleData, ?popaggData and ?popoutData are
# implemented in the rsarGWR-package. See help files.

\dontrun{

formula <- y~1+x|clusterid |long + lat

#Model fit
gwmodel<-gwlmm(formula, data = data)

#In-sample predictions
pred<-predict(gwmodel)

#Small area mean prediction for aggregated population data
predagg<-predict(gwmodel, popdata = popaggData, size = "Size")

#Small area mean prediction for unit-level population data
preddisagg<-predict(gwmodel, popdata = popoutData, popAgg = FALSE)
}
##################################################################
# Outlier-robust estimation
\dontrun{

# Model fit
rgwmodel<- rgwlmm(formula, data = sampleData)

# In-sample prediction
rpred<-predict(rgwmodel)
#Small area preditions (mean) for aggregated population data
rpredagg<-predict(rgwmodel, popdata = popaggData, size = "Size")
#Small area preditions (mean) for unit-level population data
rpreddisagg<-predict(rgwmodel, popdata = popoutData, popAgg = FALSE)

###########

# Robust model fit when sample only contains centroid information
rgwmodel<- rgwlmm(formula, data = sampleData, centroid = TRUE)

# In-sample prediction
rpred<-predict(rgwmodel)
#Small area means for aggregated population data
rpredagg<-predict(rgwmodel, popdata = popaggData, size = "Size")

}


}
\references{
Chambers, R., J. Chandra, and N. Tzavidis (2011). On bias-robust mean
squared error estimation for pseudo-linear small area estimators. Survey
Methodology 37 (2), 153 - 170.

Chambers, R., H. Chandra, N. Salvati, and N. Tzavidis (2014). Outlier
robust small area estimation. Journal of the Royal Statistical Society:
Series B 76 (1), 47 - 69.
}
\seealso{
\code{\link{rgwlmm}}, and \code{\link{gwlmm}}
}

