% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleData}
\alias{sampleData}
\title{Generated Samlpe for demonstrational purpose of the rsaeGWR package}
\format{A data frame with 200 rows and 5 variable:
\describe{
  \item{long}{Longituge}
  \item{lat}{Latitude}
  \item{clusterid}{ID variable, defining the areas}
  \item{x}{Continious explanatory variable}
  \item{y}{Continious dependend variable}...
}}
\usage{
data(sampleData)
}
\description{
Generated Samlpe for demonstrational purpose of the rsaeGWR package
}
\keyword{datasets}

